requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==21.1
django-cors-headers==3.10.1
django-registration==3.1.2
django-registration-redux==2.9
django-tinymce==3.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.0
